+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/frontend/CalciteRexNode.java
@@ -1,0 +1,21 @@
+package org.dbsp.sqlCompiler.compiler.frontend;
+
+import org.apache.calcite.rex.RexNode;
+
+public class CalciteRexNode extends CalciteObject {
+    final RexNode rexNode;
+
+    CalciteRexNode(RexNode rexNode) {
+        this.rexNode = rexNode;
+    }
+
+    @Override
+    public boolean isEmpty() {
+        return false;
+    }
+
+    @Override
+    public String toString() {
+        return this.rexNode.toString();
+    }
+}