+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [SQL] Fix bugs in parsing of KEY and FOREIGN KEY constraints
+
 ## [0.3.2] - 2023-11-10
 
 ### Fixed